Fix beaker on xenial
Since the beaker jobs are being run on xenial, we need a special nodeset
for it, otherwise beaker gives an error:

 beaker-hostgenerator was not able to use this value as input.
 Exiting with an Error.

We also want to install puppet from the Ubuntu repos rather than from
puppetlabs, since puppetlabs doesn't support puppet 3 for Xenial. For
centos we can keep the install process the same.

This patch includes additional fixes for issues that were probably
present when the tests were first merged:

- make sure the irc pip package is installed by the puppet class
- use double quotes for the channels.yaml fixture so properly render the
  newline characters
- add "access" and "global" sections to the channels.yaml fixture since
  accessbot will otherwise be unable to parse the file
- remove spec assertions about the accessbot process, because if it was
  successful it should have exited during the puppet run, and in fact if
  it didn't exit the puppet run would have failed.
- correct the group name the spec is asserting of the config file

Additionally, out of band from this patch, the nick 'accessbot-test' is
nowregistered on freenode. If it's not registered, it becomes stuck in
the irc infinite loop without performing any actions. This does not mean
it has any permissions to make any real access changes.

Change-Id: Ifd2244ae9dd212b2475f9cd6adb994bc058a4769
Depends-On: I23d0a9c0f4b6ecbb3403447adb03e845d2695366