Add mlock setting and op_channel option
Add a mode option, which is passed through to mlock to set the modes
on channels.  Individual channels can override the global setting.

Because you must be an operator in the channel you are forwarding to
when using the +f mode, add a op_channel option that pre-joins a given
channel and promotes ourselves to ops in there.

[1] https://wiki.swiftirc.net/wiki/Chanserv#Mlock

Change-Id: Ic3ea018f1dd3032a628030ed56cfd863de2f159d